Add co2 price to sector networks as marginal_cost on the `co2 atmosphere` Store. (#1897)
* Add co2 price from costs: emission_prices to sector networks

Adds the -co2_price as marginal_cost on the co2 atmosphere tracking store. So that
storing a ton costs co2_price and taking out of the atmosphere leads to a gain of
co2_price.

* fix: ignore dict-like emission prices in the electricity sector

those define co2 prices for multiple planning horizons in the myopic optimization.

* fix (docs): add co2 price updates to costs configtable

* Update release_notes.rst

* fix (docs): advise to use opts wildcard only for electricity-only runs

* fix: release notes

* fix: fix plot_balance_map with a single negative price

* fix: allow removing co2_budget with None.

ie. to remove a co2_budget set in config.default.yaml, use: ``` co2_budget: ```